ai-system/requirements.txt
openai>=1.0.0
pinecone
gradio>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
requests

PyPDF2
python-docx
pandas
openpyxl
ebooklib
beautifulsoup4
chardet
sentence-transformers
peft
transformers
torch
accelerate
bitsandbytes
langchain 
langgraph
